add voice messages

+++ app/src/main/java/com/nextcloud/talk/jobs/UploadAndShareFilesWorker.kt
@@ -41,6 +41,7 @@ import com.nextcloud.talk.utils.ApiUtils
 import com.nextcloud.talk.utils.UriUtils
 import com.nextcloud.talk.utils.bundle.BundleKeys.KEY_FILE_PATHS
 import com.nextcloud.talk.utils.bundle.BundleKeys.KEY_INTERNAL_USER_ID
+import com.nextcloud.talk.utils.bundle.BundleKeys.KEY_META_DATA
 import com.nextcloud.talk.utils.bundle.BundleKeys.KEY_ROOM_TOKEN
 import com.nextcloud.talk.utils.database.user.UserUtils
 import com.nextcloud.talk.utils.preferences.AppPreferences
@@ -88,6 +89,7 @@ class UploadAndShareFilesWorker(val context: Context, workerParameters: WorkerPa
             val sourcefiles = inputData.getStringArray(DEVICE_SOURCEFILES)
             val ncTargetpath = inputData.getString(NC_TARGETPATH)
             val roomToken = inputData.getString(ROOM_TOKEN)
+            val metaData = inputData.getString(META_DATA)
 
             checkNotNull(currentUser)
             checkNotNull(sourcefiles)
@@ -99,7 +101,7 @@ class UploadAndShareFilesWorker(val context: Context, workerParameters: WorkerPa
                 val sourcefileUri = Uri.parse(sourcefiles[index])
                 val filename = UriUtils.getFileName(sourcefileUri, context)
                 val requestBody = createRequestBody(sourcefileUri)
-                uploadFile(currentUser, ncTargetpath, filename, roomToken, requestBody, sourcefileUri)
+                uploadFile(currentUser, ncTargetpath, filename, roomToken, requestBody, sourcefileUri, metaData)
             }
         } catch (e: IllegalStateException) {
             Log.e(javaClass.simpleName, "Something went wrong when trying to upload file", e)
@@ -130,7 +132,8 @@ class UploadAndShareFilesWorker(val context: Context, workerParameters: WorkerPa
         filename: String,
         roomToken: String?,
         requestBody: RequestBody?,
-        sourcefileUri: Uri
+        sourcefileUri: Uri,
+        metaData: String?
     ) {
         ncApi.uploadFile(
             ApiUtils.getCredentials(currentUser.username, currentUser.token),
@@ -151,7 +154,7 @@ class UploadAndShareFilesWorker(val context: Context, workerParameters: WorkerPa
                 }
 
                 override fun onComplete() {
-                    shareFile(roomToken, currentUser, ncTargetpath, filename)
+                    shareFile(roomToken, currentUser, ncTargetpath, filename, metaData)
                     copyFileToCache(sourcefileUri, filename)
                 }
             })
@@ -159,17 +162,28 @@ class UploadAndShareFilesWorker(val context: Context, workerParameters: WorkerPa
 
     private fun copyFileToCache(sourceFileUri: Uri, filename: String) {
         val cachedFile = File(context.cacheDir, filename)
-        val outputStream = FileOutputStream(cachedFile)
-        val inputStream: InputStream = context.contentResolver.openInputStream(sourceFileUri)!!
 
-        inputStream.use { input ->
-            outputStream.use { output ->
-                input.copyTo(output)
+        if (cachedFile.exists()) {
+            Log.d(TAG, "file is already in cache")
+        } else {
+            val outputStream = FileOutputStream(cachedFile)
+            val inputStream: InputStream = context.contentResolver.openInputStream(sourceFileUri)!!
+
+            inputStream.use { input ->
+                outputStream.use { output ->
+                    input.copyTo(output)
+                }
             }
         }
     }
 
-    private fun shareFile(roomToken: String?, currentUser: UserEntity, ncTargetpath: String?, filename: String?) {
+    private fun shareFile(
+        roomToken: String?,
+        currentUser: UserEntity,
+        ncTargetpath: String?,
+        filename: String?,
+        metaData: String?) {
+
         val paths: MutableList<String> = ArrayList()
         paths.add("$ncTargetpath/$filename")
 
@@ -177,6 +191,7 @@ class UploadAndShareFilesWorker(val context: Context, workerParameters: WorkerPa
             .putLong(KEY_INTERNAL_USER_ID, currentUser.id)
             .putString(KEY_ROOM_TOKEN, roomToken)
             .putStringArray(KEY_FILE_PATHS, paths.toTypedArray())
+            .putString(KEY_META_DATA, metaData)
             .build()
         val shareWorker = OneTimeWorkRequest.Builder(ShareOperationWorker::class.java)
             .setInputData(data)
@@ -190,6 +205,7 @@ class UploadAndShareFilesWorker(val context: Context, workerParameters: WorkerPa
         const val DEVICE_SOURCEFILES = "DEVICE_SOURCEFILES"
         const val NC_TARGETPATH = "NC_TARGETPATH"
         const val ROOM_TOKEN = "ROOM_TOKEN"
+        const val META_DATA = "META_DATA"
 
         fun isStoragePermissionGranted(context: Context): Boolean {
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
